requirements.txt
fastapi==0.109.2
alembic==1.13.1
sqlalchemy==2.0.27
uvicorn==0.27.1
psycopg-binary==3.1.18
python-dotenv==1.0.1
pydantic-settings==2.2.0
psycopg2-binary==2.9.9
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2
